master: Validate inputs and responses with their proper status codes

+++ app/Http/Controllers/StoreController.php
@@ -42,17 +42,32 @@ class StoreController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
-        $request->validate(['name' => 'required|string|max:255']);
+        $request->validate([
+            'name' => 'bail|required|string|max:255',
+            'products' => 'array',
+            'products.*.name' => 'required|string|max:255',
+            'products.*.quantity' => 'required|integer|min:0',
+        ]);
+
         $store = Store::factory()->create(['name' => $request->name]);
 
-        foreach ($request->products as $productRequest) {
-            $product = Product::factory()->create(['name' => $productRequest['name']]);
+        foreach ($request->products ?? [] as $productRequest) {
+            $product = Product::updateOrCreate(
+                ['id' => $productRequest['id'] ?? null],
+                ['name' => $productRequest['name']]
+            );
 
-            ProductStore::factory()->create([
-                'product_id' => $product->id,
-                'store_id' => $store->id,
-                'quantity' => $productRequest['quantity'] ?? 0,
-            ]);
+            ProductStore::updateOrCreate(
+                [
+                    'product_id' => $product->id,
+                    'store_id' => $store->id
+                ],
+                [
+                    'product_id' => $product->id,
+                    'store_id' => $store->id,
+                    'quantity' => $productRequest['quantity'] ?? 0,
+                ]
+            );
         }
 
         return response()->json(['message' => 'Store created successfully'], 201);
@@ -61,10 +76,14 @@ class StoreController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(int $id)
     {
         $data = [];
         $store = Store::with('products')->find($id);
+
+        if (!$store) {
+            return response()->json(['message' => 'Store not found'], 404);
+        }
 
         $data[] = [
             'id' => $store->id,
@@ -84,21 +103,42 @@ class StoreController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id): JsonResponse
+    public function update(Request $request, int $id): JsonResponse
     {
-        $request->validate(['name' => 'required|string|max:255']);
+        if (null === Store::with('products')->find($id)) {
+            return response()->json(['message' => 'Store not found'], 404);
+        }
+
+        $request->validate([
+            'name' => 'required|string|max:255',
+            'products' => 'array',
+            'products.*.name' => 'required|string|max:255',
+            'products.*.quantity' => 'required|integer|min:0',
+        ]);
+
         Store::with('products')->find($id)->update(['name' => $request->name]);
 
-        foreach ($request->products as $productRequest) {
-            Product::find($productRequest['id'])->update([
-               'name' => $productRequest['name'],
+        foreach ($request->products ?? [] as $productRequest) {
+            $product = Product::updateOrCreate(
+                ['id' => $productRequest['id'] ?? null],
+                ['name' => $productRequest['name'],
             ]);
 
-            ProductStore::where('product_id', $productRequest['id'])
+            $productStore = ProductStore::where('product_id', $product->id)
                 ->where('store_id', $id)
-                ->update([
-                    'quantity' => $productRequest['quantity'] ?? 0,
+                ->first();
+
+            if (null === $productStore) {
+                ProductStore::create([
+                    'product_id' => $product->id,
+                    'store_id' => $id,
+                    'quantity' => $productRequest['quantity']
                 ]);
+            } else {
+                // @TODO fix composite primary key of ProductStore in order to update
+                $productStore->quantity = $productRequest['quantity'];
+                $productStore->save();
+            }
         }
 
         return response()->json(['message' => 'Store updated successfully'], 204);
@@ -107,8 +147,12 @@ class StoreController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(int $id)
     {
+        if (null === Store::with('products')->find($id)) {
+            return response()->json(['message' => 'Store not found'], 404);
+        }
+
         Store::with('products')->find($id)->delete();
 
         return response()->json(['message' => 'Store deleted successfully'], 204);
@@ -116,11 +160,16 @@ class StoreController extends Controller
 
     public function sellProductStore(Request $request, int $storeId, int $productId): JsonResponse
     {
-        $request->validate(['quantity' => 'required|integer|min:1']);
-
-        $currentStock = ProductStore::where('product_id', $productId)
+        $productStore = ProductStore::where('product_id', $productId)
             ->where('store_id', $storeId)
-            ->value('quantity');
+            ->first();
+
+        if (null === $productStore) {
+            return response()->json(['message' => 'Product not found in store'], 404);
+        }
+
+        $request->validate(['quantity' => 'required|integer|min:1']);
+        $currentStock = $productStore->quantity;
 
         if ($currentStock < $request->quantity) {
             return response()->json(['message' => "Insufficient stock ($currentStock units), sell cannot be made."], 400);
@@ -141,7 +190,7 @@ class StoreController extends Controller
         if ($currentStock <= self::LOW_STOCK_THRESHOLD) {
             $data['message'] .= " Low stock ($currentStock units), please restock soon.";
 
-            return response()->json($data, 200);
+            return response()->json($data);
         }
 
         return response()->json(null, 204);
